+++ src/lib/pupiq.php
@@ -9,7 +9,7 @@ defined("PUPIQ_DEFAULT_WATERMARK_DEFINITION") || define("PUPIQ_DEFAULT_WATERMARK
 
 class Pupiq {
 
-	const VERSION = "1.12.1";
+	const VERSION = "1.13";
 
 	protected $_api_key = "";
 
@@ -30,8 +30,8 @@ class Pupiq {
 
 	protected $_lang = PUPIQ_LANG;
 
-	static protected $_SupportedImageFormats = array("jpg","png","svg");
-	static protected $_ImageFormatsSupportingTransparency = array("png","svg");
+	static protected $_SupportedImageFormats = array("jpg","png","webp","svg");
+	static protected $_ImageFormatsSupportingTransparency = array("png","svg","webp");
 
 	function __construct($url_or_api_key = "",$api_key = null){
 		$url = "";
@@ -60,6 +60,13 @@ class Pupiq {
 		}
 	}
 
+	/**
+	 *
+	 *	$image = Pupiq::CreateImage("https://example.com/images/flower.jpg");
+	 *
+	 *	$image = Pupiq::CreateImage("/path/to/flower.jpg");
+	 *	$image = Pupiq::CreateImage(["path" => "/path/to/temp_file", "name" => "flower.jpg"]);
+	 */
 	static function CreateImage($url_or_filename,&$err_msg = ""){
 		$pupiq = new Pupiq();
 
@@ -69,14 +76,27 @@ class Pupiq {
 		$options = array(
 			"acceptable_error_codes" => array(400,403),
 		);
-		$file = null;
+
+		$url = null;
+		$file = array();
+
+		if(is_array($url_or_filename)){
+			$file = $url_or_filename + array(
+				"path" => null, // "/path/to/a/tempory_file"
+				"name" => null, // "photo.jpg"
+			);
+		}elseif(preg_match('/^https?:/',$url_or_filename)){
+			$url = $url_or_filename;
+		}else{
+			$file["path"] = $url_or_filename;
+		}
 
 		$adf = new ApiDataFetcher(PUPIQ_API_URL);
-		if(preg_match('/^https?:/',$url_or_filename)){
-			$params["url"] = $url_or_filename;
+		if($url){
+			$params["url"] = $url;
 			$data = $adf->post("images/create_new",$params,$options);
 		}else{
-			$data = $adf->postFile("images/create_new",$url_or_filename,$params,$options);
+			$data = $adf->postFile("images/create_new",$file,$params,$options);
 		}
 
 		if(!$data){
@@ -369,12 +389,12 @@ class Pupiq {
 	}
 
 	/**
-	 * echo $pupiq->getTransformation($force_suffix); // "800x600", $force_suffix=="jpeg"
+	 * echo $pupiq->getTransformation($force_suffix); // "800x600", $force_suffix=="jpg"
 	 * echo $pupiq->getTransformation($force_suffix); // "800x600,transparent", $force_suffix=="png"
 	 */
 	function getTransformation(&$force_suffix = null){
 		$force_suffix = $this->_suffix;
-		if($this->_format){
+		if($this->_format && in_array($this->_format,array("jpg","png","webp")) && in_array($this->_suffix,array("jpg","png","webp"))){
 			$force_suffix = $this->_format;
 		}
 		if(!$this->_transformation_string){