+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+Change Log
+==========
+
+All notable changes to PupiqClient will be documented in this file.
+
+## [1.13] - 2021-02-08
+
+- Added AsyncPupiqAttachmentField (package atk14/async-file-field needs to be installed in order to use AsyncPupiqAttachmentField)
+
+## [1.12.4] - 2021-02-07
+
+- Dependency updated
+
+## [1.12.3] - 2020-08-19
+
+- Fix: svg image can't be converted into png or jpg (the image stays as svg)
+
+## [1.12.2] - 2020-05-08
+
+- In PupiqImageInput, thumbnail transformation corrected
+
 ## [1.12.1] - 2020-04-12
 
 - "Transparent or colorful background" option fixed for SVG images